Removed zcash mod and renamed multi_sig mod
