+++ mithril-core/src/lib.rs
@@ -8,6 +8,6 @@ pub mod key_reg;
 mod merkle_tree;
 pub mod stm;
 
-mod multi_sig_blast;
+mod multi_sig;
 
 pub use crate::error::AggregationError;
