+++ mithril-core/src/stm.rs
@@ -109,7 +109,7 @@ use crate::dense_mapping::ev_lt_phi;
 use crate::error::{AggregationError, RegisterError, StmSignatureError};
 use crate::key_reg::ClosedKeyReg;
 use crate::merkle_tree::{MTLeaf, MerkleTreeCommitment, Path};
-use crate::multi_sig_blast::{Signature, SigningKey, VerificationKey, VerificationKeyPoP};
+use crate::multi_sig::{Signature, SigningKey, VerificationKey, VerificationKeyPoP};
 use digest::{Digest, FixedOutput};
 use rand_core::{CryptoRng, RngCore};
 use serde::{Deserialize, Serialize};
